Fix compatible with other plugins which overwrites projects_helper
