+++ lib/redmine_messenger/patches/projects_helper_patch.rb
@@ -1,7 +1,5 @@
 # Redmine Messenger plugin for Redmine
 
-require_dependency 'projects_helper'
-
 module RedmineMessenger
   module Patches
     module ProjectsHelperPatch
@@ -13,12 +11,6 @@ module RedmineMessenger
       end
 
       module InstanceMethods
-        def project_messenger_options(active)
-          options_for_select({ l(:label_messenger_settings_default) => 0,
-                               l(:label_messenger_settings_disabled) => 1,
-                               l(:label_messenger_settings_enabled) => 2 }, active)
-        end
-
         def project_settings_tabs_with_messenger
           tabs = project_settings_tabs_without_messenger
           action = { name: 'messenger',
@@ -29,14 +21,6 @@ module RedmineMessenger
 
           tabs << action if User.current.allowed_to?(:manage_messenger, @project)
           tabs
-        end
-
-        def project_setting_messenger_default_value(value)
-          if Messenger.default_project_setting(@project, value)
-            l(:label_messenger_settings_enabled)
-          else
-            l(:label_messenger_settings_disabled)
-          end
         end
       end
     end
